requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.51.0
python-multipart==0.0.9
pillow==11.0.0
pymupdf==1.24.10
pydantic-settings==2.6.0
httpx==0.27.2
rapidocr-onnxruntime>=1.3.0